+++ src/sinks/util/buffer.rs
@@ -1,3 +1,4 @@
+use super::{batch::Batch, partition::Partition};
 use flate2::write::GzEncoder;
 use serde::{Deserialize, Serialize};
 use std::io::Write;
@@ -63,7 +64,7 @@ impl Buffer {
     }
 }
 
-impl super::batch::Batch for Buffer {
+impl Batch for Buffer {
     type Input = Vec<u8>;
     type Output = Vec<u8>;
 
@@ -103,6 +104,83 @@ impl super::batch::Batch for Buffer {
 
     fn num_items(&self) -> usize {
         self.num_items
+    }
+}
+
+#[derive(Debug)]
+pub struct PartitionBuffer<T, K> {
+    inner: T,
+    key: Option<K>,
+}
+
+#[derive(Debug, Clone)]
+pub struct PartitionInnerBuffer<T, K> {
+    pub(self) inner: T,
+    key: K,
+}
+
+impl<T, K> PartitionBuffer<T, K> {
+    pub fn new(inner: T) -> Self {
+        Self { inner, key: None }
+    }
+}
+
+impl<T, K> Batch for PartitionBuffer<T, K>
+where
+    T: Batch,
+    K: Clone,
+{
+    type Input = PartitionInnerBuffer<T::Input, K>;
+    type Output = PartitionInnerBuffer<T::Output, K>;
+
+    fn len(&self) -> usize {
+        self.inner.len()
+    }
+
+    fn push(&mut self, item: Self::Input) {
+        let partition = item.partition();
+        self.key = Some(partition);
+        self.inner.push(item.inner)
+    }
+
+    fn is_empty(&self) -> bool {
+        self.inner.is_empty()
+    }
+
+    fn fresh(&self) -> Self {
+        Self {
+            inner: self.inner.fresh(),
+            key: None,
+        }
+    }
+
+    fn finish(mut self) -> Self::Output {
+        let key = self.key.take().unwrap();
+        let inner = self.inner.finish();
+        PartitionInnerBuffer { inner, key }
+    }
+
+    fn num_items(&self) -> usize {
+        self.inner.num_items()
+    }
+}
+
+impl<T, K> PartitionInnerBuffer<T, K> {
+    pub fn new(inner: T, key: K) -> Self {
+        Self { inner, key }
+    }
+
+    pub fn into_parts(self) -> (T, K) {
+        (self.inner, self.key)
+    }
+}
+
+impl<T, K> Partition<K> for PartitionInnerBuffer<T, K>
+where
+    K: Clone,
+{
+    fn partition(&self) -> K {
+        self.key.clone()
     }
 }