+++ CHANGELOG.md
@@ -10,15 +10,19 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
-- aws_s3: Add `filename_extension` options
+- aws_s3: Add `filename_extension` options.
+- aws_cloudwatch_logs: `stream_name` now accepts `{{key}}` synatx for extracting values from events.
 
 ### Changed
+
+- aws_cloudwatch_logs: Now partitions events by `log_group`/`log_stream`.
 
 ### Deprecated
 
 ### Fixed
 
 - aws_s3: Fixed #517 and trailing slash issues with the generated key.
+- aws_cloudwatch_logs: Fixes #586 and now dynamically creates streams if they do not exist.
 
 ### Removed